web/requirements.txt
Django>=4.2
psycopg2-binary
django-crontab
pytz
Pillow>=11
django-modeltranslation>=0.17
cookiecutter>=2.0


